pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "robotwin-core"
version = "1.0.0"
description = "RoboTwin Platform - Robot Core Control System for Ubuntu Mini PC"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "RoboTwin Team" }
]

dependencies = [
    # WebSocket & Communication
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    "paho-mqtt>=2.0.0",
    
    # ROS2 Python Client (install via apt for full ROS integration)
    # rclpy - installed via ROS2 installation
    
    # Hardware & GPIO
    "pyserial>=3.5",
    "smbus2>=0.4.3",
    "RPi.GPIO>=0.7.1; platform_system != 'Darwin'",
    "gpiozero>=2.0; platform_system != 'Darwin'",
    
    # Computer Vision & Camera
    "opencv-python>=4.9.0",
    "opencv-contrib-python>=4.9.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    
    # Math & Control
    "scipy>=1.11.0",
    "filterpy>=1.4.5",
    "transforms3d>=0.4.1",
    
    # AI/ML (optional, for advanced features)
    "torch>=2.0.0; extra == 'ml'",
    "ultralytics>=8.0.0; extra == 'ml'",
    
    # Utilities
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.0.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "aiofiles>=23.0.0",
    
    # Monitoring
    "psutil>=5.9.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.6.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]
ml = [
    "torch>=2.0.0",
    "ultralytics>=8.0.0",
    "onnxruntime>=1.16.0",
]
ros = [
    # ROS2 dependencies - install via apt
]

[project.scripts]
robotwin = "robotwin.cli:main"
robot-server = "robotwin.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/robotwin"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "PT"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
